Update Jupyter Notebook to include RAG model and model index
Modified the RAG-with-LLaMa.ipynb file to include loading of the model ID and model index of the Hugging Face's RAG model and Pinecone's model, respectively. It also includes the usage of the HuggingFaceEmbeddings class, which maps inputs to a vector space. The section of Initializing the Hugging Face Embedding Pipeline has been restructured and detailed out.